`alias` is still fine to share
